refactor upstream validator
